{"schema":"decodex/commit/1","summary":"Split control-plane facade modules","authority":"manual"}
